refactor: move tests out of src
